% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asP01.R
\name{as.P01}
\alias{as.P01}
\title{GF# to P01}
\usage{
as.P01(gf3)
}
\arguments{
\item{gf3}{a gf3 standard code paramater}
}
\value{
a matching P01 value with units and standard name (if applicable)
}
\description{
Use this function to map gf3 codes to P01 codes for exporting to netCDF
}
