% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd_nc.R
\name{ctd_nc}
\alias{ctd_nc}
\title{CTD netCDF template}
\usage{
ctd_nc(obj, upcast = NULL, metadata, filename = NULL)
}
\arguments{
\item{obj}{an odf object from oce which contains mctd data}

\item{metadata}{a csv file following the standard template which includes all
necessary metadata}

\item{filename}{the desired name for the netCDF file produced, if left NULL
the default will conform to BIO naming conventions}
}
\value{
netCDF file with variables temperature, conductivity, pressure, sigma
  theta, oxygen, salinity, station, latitude, longitude, oxygen voltage,
  fluoresence, par, scan, scan2 and QC flag variables for each data variable
}
\description{
CTD netCDF template
}
\examples{
file <- list.files('.', pattern = "CTD*...*.ODF")
metadata <- ('CTD_SAMPLE_METADATA.csv')
obj <- read.odf(file[1])
upcast <- read.odf(file[2])
mcm_nc(obj, upcast, metadata)
}
