% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mctd_nc.R
\name{mctd_nc}
\alias{mctd_nc}
\title{Moored CTD netCDF template}
\usage{
mctd_nc(obj, metadata, filename = NULL)
}
\arguments{
\item{obj}{an odf object from oce which contains mctd data}

\item{metadata}{a csv file following the standard template which includes all
necessary metadata}

\item{filename}{the desired name for the netCDF file produced, if left NULL
the default will conform to BIO naming conventions}
}
\value{
netCDF file with a maximum of 12 variables
}
\description{
Moored CTD netCDF template
}
\examples{
file <- list.files('.', pattern = "MCTD*...*.ODF")
obj <- read.odf(file)
metadata <- 'MCTD_SAMPLE_METADATA.csv'
mctd_nc(obj, metadata)

}
