% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtr_nc.R
\name{mtr_nc}
\alias{mtr_nc}
\title{Moored temperature recorder netCDF template}
\usage{
mtr_nc(obj, metadata, filename = NULL)
}
\arguments{
\item{obj}{an odf object from oce which contains temperature data}

\item{metadata}{a csv file following the standard template which includes all
necessary metadata}

\item{filename}{the desired name for the netCDF file produced, if left NULL
the default will conform to BIO naming conventions}
}
\value{
netCDF file with up to 12 variables
}
\description{
Moored temperature recorder netCDF template
}
\examples{
file <- list.files('.', pattern = "MTR*...*.ODF")
obj <- read.odf(file)
metadata <- "MTR_SAMPLE_METADATA.csv"
mtr_nc(obj, metadata)
}
