% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odf_check.R
\name{odf_check}
\alias{odf_check}
\title{Check ODF metadata}
\usage{
odf_check(obj, print = TRUE)
}
\arguments{
\item{obj}{an odf object (oce::read.odf())}

\item{print}{TRUE or FALSE, TRUE will cause errors to be displayed at command
line, FALSE will sink errors into text document}
}
\value{
will print any issues with metadata to command line, if nothing
  prints then all metadata is intact
}
\description{
Check that an ODF object has all the metadata required to build a complete netCDF file
}
