% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readodfcm.R
\name{read.odf.cm}
\alias{read.odf.cm}
\title{MCM odf 2 oce}
\usage{
read.odf.cm(filename)
}
\arguments{
\item{filename}{odf file containing Moored current metre (MCM) data}
}
\value{
an oce object, cm - class containing all the data from the odf file
}
\description{
MCM odf 2 oce
}
