% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readodfmctd.R
\name{read.odf.mctd}
\alias{read.odf.mctd}
\title{mctd odf 2 oce}
\usage{
read.odf.mctd(filename)
}
\arguments{
\item{filename}{odf file containing moored CTD data (MCTD)}
}
\value{
oce object, ctd class containing all the data and metadata from the ODF
}
\description{
mctd odf 2 oce
}
